Se modifican los filtros y se separan los botones
